Build system template
This imports the essential elements of the build system from usdpaa, and
puts a template directory structure in place to accept the upcoming
user-space port of the layerscape qbman driver.

The header files, driver implementation, and test app are just "hello
world" place-holders to check that the tool-chain and build-system are
functioning.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>