qbman: Workaround for the DQRR reset bug.
On portal reset, the indices are reset but the actual ring entries may have
stale data that isn't cleared. This means for the first run around the ring, we
can't use valid-bits when polling ring entries until we're sure each entry has
already been used once. So we read the PI index, and force an invalidate and
reread of the ring entry once PI indicates that h/w has something at that
location. This workaround disables itself once it has wrapped around the ring
once.
This workaround works for ERR008838.

Change-Id: If4a01820a23f569d4264c0032933b9c9f9b4ee3b
Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/30502
Reviewed-by: Roy Pledge <roy.pledge@freescale.com>