qbman: Make token tracking internal
Previousy the QBMan portal code required the caller to properly set and
check for a token value used by the driver to detect when the QMan
hardware had completed a dequeue.  This patch simplifes the driver
interface by internally dealing with token values.  The driver will now
set the token value to 0 once it has dequeued a frame while a token
value of 1 indicates the HW has completed the dequeue but SW has not
consumed the frame yet.

Signed-off-by: Roy Pledge <Roy.Pledge@freescale.com>
Change-Id: I6f7672ed7b4d2a9b8988ae39aae900dda31f4786
Reviewed-on: http://git.am.freescale.net:8181/34868
Reviewed-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Reviewed-by: Haiying Wang <Haiying.Wang@freescale.com>
Tested-by: Haiying Wang <Haiying.Wang@freescale.com>