qbman: rename qbman_dq_entry to qbman_result
Currently qbman_dq_entry is used for both dq result in dqrr
and memory, and notifications in dqrr and memory. It doesn't
make sense to have dq_entry in name for those notifications
which have nothing to do with dq. So we rename this as
qbman_result which is meaningful for both cases.

Signed-off-by: Haiying Wang <Haiying.wang@freescale.com>