driver: Fix BUG_ON related unused var warnings
Removed the verb variable in driver/qbman_portal.c and
driver/qbman_debug.c functions that was used in BUG_ON() macro. Placed
the call that populated the verb variable directly inside the BUG_ON()
macro. This way we avoid the compiler warning if BUG_ON() macro is
compiled out. Readability of the code is slightly diminished

Signed-off-by: Ahmed Mansour <Ahmed.Mansour@freescale.com>