driver/qbman_portal.h: Fix signed unsigned compare
Eliminate ambiguity in unsigned signed comparison. The intention was
to compare unsigned with unsigned. The compiler saw the right argument
of the comparison as signed. This could cause issues in an edge case
where the right argument would be seen as a negative number instead of
a large positive integer.

Signed-off-by: Ahmed Mansour <Ahmed.Mansour@freescale.com>