driver/portal: Update discrete consumption logic
Update logic/macros that handle the setting of the discrete
consumption in the enqueue descriptor. The logic worked previously for
ls2085, but not for ls2088. The size of the ring was defined and used
in multiple places which made updates difficult.  This patch
consolidates the information to a smaller number of locations. The
same binary can now be run on either ls2088, ls2085 or others

Signed-off-by: Ahmed Mansour <Ahmed.Mansour@freescale.com>